requirements.txt
asgiref==3.8.1
Django==5.0.7
django-rest-framework==0.1.0
djangorestframework==3.15.1
mysqlclient==2.2.4
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
djangorestframework-simplejwt==5.3.1
whitenoise
python-decouple==3.8
django-tidb>=5.0.0,<5.1.0